chore: useHasDynamicHooks
