+++ apps/web/src/views/SwapSimplify/InfinitySwap/TradingFee.tsx
@@ -1,15 +1,12 @@
-import { findHook, HOOK_CATEGORY } from '@pancakeswap/infinity-sdk'
 import { useTranslation } from '@pancakeswap/localization'
-import { OrderType, PriceOrder } from '@pancakeswap/price-api-sdk'
-import { PoolType } from '@pancakeswap/smart-router'
+import { PriceOrder } from '@pancakeswap/price-api-sdk'
 import { FlexGap, SkeletonV2, Text } from '@pancakeswap/uikit'
 import { formatAmount } from '@pancakeswap/utils/formatFractions'
 import { memo, useMemo } from 'react'
-import { Address } from 'viem'
-import { isAddress } from 'viem/utils'
 import { isXOrder } from 'views/Swap/utils'
 import { useIsWrapping, useSlippageAdjustedAmounts } from '../../Swap/V3Swap/hooks'
 import { computeTradePriceBreakdown } from '../../Swap/V3Swap/utils/exchange'
+import { useHasDynamicHook } from '../hooks/useHasDynamicHook'
 
 interface TradingFeeProps {
   loaded: boolean
@@ -24,21 +21,7 @@ export const TradingFee: React.FC<TradingFeeProps> = memo(({ order, loaded }) =>
     [order],
   )
 
-  const containsDynamicHooks = useMemo(() => {
-    if (!order?.trade || order.type !== OrderType.PCS_CLASSIC) return false
-    const chainId = order.trade.inputAmount.currency.chainId
-    const pools = order.trade.routes
-      .map((r) => r.pools)
-      .flat()
-      .filter((pool) => pool.type === PoolType.InfinityBIN || pool.type === PoolType.InfinityCL)
-    if (pools.length === 0) return false
-
-    const hooks = pools.map((pool) => pool.hooks).filter((hook) => hook && isAddress(hook)) as Address[]
-
-    return hooks.some((hook) => {
-      return findHook(hook, chainId)?.category?.includes(HOOK_CATEGORY.DynamicFees) ?? false
-    })
-  }, [order?.trade])
+  const hasDynamicHooks = useHasDynamicHook(order)
   const isWrapping = useIsWrapping()
 
   if (isWrapping || !order || !order.trade || !slippageAdjustedAmounts) {
@@ -57,10 +40,9 @@ export const TradingFee: React.FC<TradingFeeProps> = memo(({ order, loaded }) =>
             0 {inputAmount?.currency?.symbol}
           </Text>
         ) : (
-          <Text color="textSubtle" fontSize="14px">{`${containsDynamicHooks ? '~' : ''}${formatAmount(
-            lpFeeAmount,
-            4,
-          )} ${inputAmount?.currency?.symbol}`}</Text>
+          <Text color="textSubtle" fontSize="14px">{`${hasDynamicHooks ? '~' : ''}${formatAmount(lpFeeAmount, 4)} ${
+            inputAmount?.currency?.symbol
+          }`}</Text>
         )}
       </SkeletonV2>
     </FlexGap>
